Fixed the final result set re-sizing
Before, the final result set would return with NIL values in the unused-allocated spaces of the tuple. I updated the result set to be an array of tuples rather than a tuple of tuples so the client could append to it to add rows. Also, the creation of the final Python list only includes rows and not the NIL values. Looks like there still is an issue with the actual resizing of the C array though if the result set requires resizing due to a lot of rows being returned.