Wrote a method for reading a file into memory.
Currently the read_file method reads the whole file into memory and returns it to the user as a list of tuples. This may be adjusted later to attach the file in memory to the class instance as an attribute, but for now this will work. There are a few known issues here as well: 1. if the file path isn't right, it will return a segmentation fault when trying to read the file. 2. the last row seems to be an empty tuple, not sure why this is being read in like that but my guess is the parseRow is hitting the line break and then the next row starts and thats when we hit the EOF. 3. The headers are being included in the file, need to add a check to make sure we aren't reading in the headers if headers_exist is True.